+++ boot-core/src/contract.rs
@@ -3,43 +3,28 @@ use std::path::Path;
 use std::{
     cell::RefCell,
     fmt::{self, Debug},
-    marker::PhantomData,
     rc::Rc,
 };
 
+use crate::BootEnvironment;
 use crate::{
-    error::BootError,
-    index_response::IndexResponse,
-    state::StateInterface,
-    tx_handler::{TxHandler, TxResponse},
+    error::BootError, index_response::IndexResponse, state::StateInterface, tx_handler::TxResponse,
 };
 use cosmwasm_std::{Addr, Coin, CustomQuery, Empty};
 use cw_multi_test::Contract as TestContract;
 use schemars::JsonSchema;
 use serde::{de::DeserializeOwned, Serialize};
 
+#[allow(unused)]
 pub type StateReference<S> = Rc<RefCell<S>>;
 /// An instance of a contract. Contains references to the execution environment (chain) and a local state (state)
 /// The state is used to store contract addresses/code-ids
-pub struct Contract<
-    Chain: TxHandler,
-    E: Serialize + Debug,
-    I: Serialize + Debug,
-    Q: Serialize,
-    M: Serialize,
-> where
-    TxResponse<Chain>: IndexResponse,
-{
+pub struct Contract<Chain: BootEnvironment> {
     /// ID of the contract, used to retrieve addr/code-id
     pub id: String,
     source: ContractCodeReference,
     /// chain object that handles tx execution and queries.
     chain: Chain,
-    /// Indicate the type of executemsg
-    _execute_msg: PhantomData<E>,
-    _instantiate_msg: PhantomData<I>,
-    _query_msg: PhantomData<Q>,
-    _migrate_msg: PhantomData<M>,
 }
 
 #[derive(Default)]
@@ -59,9 +44,10 @@ where
 {
     /// CHecks the environment for the wasm dir configuration and returns the path to the wasm file
     pub fn get_wasm_code_path(&self) -> Result<String, BootError> {
-        let wasm_code_path = self.wasm_code_path.as_ref().ok_or_else( || BootError::StdErr(
-            "Wasm file is required to determine hash.".into(),
-        ))?;
+        let wasm_code_path = self
+            .wasm_code_path
+            .as_ref()
+            .ok_or_else(|| BootError::StdErr("Wasm file is required to determine hash.".into()))?;
 
         let wasm_code_path = if wasm_code_path.contains(".wasm") {
             wasm_code_path.to_string()
@@ -83,25 +69,12 @@ where
 }
 
 /// Expose chain and state function to call them on the contract
-impl<
-        Chain: TxHandler + Clone,
-        E: Serialize + Debug,
-        I: Serialize + Debug,
-        Q: Serialize + Debug,
-        M: Serialize + Debug,
-    > Contract<Chain, E, I, Q, M>
-where
-    TxResponse<Chain>: IndexResponse,
-{
+impl<Chain: BootEnvironment + Clone> Contract<Chain> {
     pub fn new(id: impl ToString, chain: &Chain) -> Self {
         Contract {
             id: id.to_string(),
             chain: chain.clone(),
             source: ContractCodeReference::default(),
-            _execute_msg: PhantomData,
-            _instantiate_msg: PhantomData,
-            _query_msg: PhantomData,
-            _migrate_msg: PhantomData,
         }
     }
 
@@ -128,13 +101,17 @@ where
     }
 
     // Chain interfaces
-    pub fn execute(&self, msg: &E, coins: Option<&[Coin]>) -> Result<TxResponse<Chain>, BootError> {
+    pub fn execute<E: Serialize + Debug>(
+        &self,
+        msg: &E,
+        coins: Option<&[Coin]>,
+    ) -> Result<TxResponse<Chain>, BootError> {
         log::info!("executing {}", self.id);
         self.chain
             .execute(msg, coins.unwrap_or(&[]), &self.address()?)
     }
 
-    pub fn instantiate(
+    pub fn instantiate<I: Serialize + Debug>(
         &self,
         msg: &I,
         admin: Option<&Addr>,
@@ -164,12 +141,15 @@ where
         Ok(resp)
     }
 
-    pub fn query<T: Serialize + DeserializeOwned>(&self, query_msg: &Q) -> Result<T, BootError> {
+    pub fn query<Q: Serialize + Debug, T: Serialize + DeserializeOwned>(
+        &self,
+        query_msg: &Q,
+    ) -> Result<T, BootError> {
         log::debug!("Querying {:#?} on {}", query_msg, self.address()?);
         self.chain.query(query_msg, &self.address()?)
     }
 
-    pub fn migrate(
+    pub fn migrate<M: Serialize + Debug>(
         &self,
         migrate_msg: &M,
         new_code_id: u64,
