+++ boot-core/src/daemon/state.rs
