+++ boot-core/src/keys/mod.rs
@@ -1,3 +1,4 @@
+#![allow(unused)]
 pub mod private;
 pub mod public;
 pub mod signature;
