+++ boot-core/src/lib.rs
@@ -1,38 +1,25 @@
-#![allow(dead_code)]
 mod contract;
-mod data_structures;
-pub(crate) mod environments;
-pub mod error;
-pub mod helpers;
+mod daemon;
+mod error;
+mod helpers;
 mod index_response;
+///
+pub mod interface;
 mod keys;
-mod multisig;
+mod mock;
 pub mod networks;
-pub mod old_traits;
-pub mod sender;
-pub mod state;
+pub mod prelude;
+mod state;
 mod tx_handler;
 
 // pub mod traits;
 pub use contract::Contract;
-pub use data_structures::{
-    daemon_state::{ChainInfo, DaemonState, NetworkInfo, NetworkKind},
-    mock_state::MockState,
-    tx_resp::CosmTxResponse,
-};
-pub use environments::{
-    daemon::Daemon, instantiate_custom_mock_env, instantiate_daemon_env,
-    instantiate_default_mock_env, mock_chain::Mock,
-};
+
 pub use error::BootError;
 pub use helpers::get_env_vars;
 pub use index_response::IndexResponse;
 pub use tx_handler::{TxHandler, TxResponse};
 
-#[macro_use]
-extern crate lazy_static;
-
-use data_structures::{client_types, core_types};
 /// Signals a supported execution environment
 pub trait BootEnvironment: TxHandler + Clone {}
