+++ boot-core/src/mock/mock_state.rs
