+++ boot-core/src/networks/mod.rs
@@ -1,2 +1,4 @@
-pub mod juno;
-pub mod osmosis;
+mod juno;
+mod osmosis;
+pub use juno::{JUNO_1, JUNO_DAEMON, UNI_5};
+pub use osmosis::{OSMO_4, OSMO_DAEMON};
