+++ boot-core/src/networks/osmosis.rs
@@ -1,7 +1,4 @@
-use crate::{
-    data_structures::daemon_state::{ChainInfo, NetworkInfo},
-    NetworkKind,
-};
+use crate::daemon::state::{ChainInfo, NetworkInfo, NetworkKind};
 
 pub const OSMO_CHAIN: ChainInfo = ChainInfo {
     chain_id: "osmosis",
